style: auth모듈 추가하면서 파일구조 및 도메인 변경
