+++ src/modules/auth/authException/NicknameAlreadyExistsException.ts
