--- src/modules/user/dto/user-login-response.dto.ts
@@ -1,4 +1,0 @@
-export class UserLoginResponseDto {
-    message: string;
-    userId: number; 
-  }
