--- src/modules/user/dto/user-login.dto.ts
@@ -1,13 +1,0 @@
-import { IsNotEmpty, IsString } from "class-validator";
-
-export class LoginUserDto {
-    
-    @IsString()
-    @IsNotEmpty()
-    accountID: string;
-
-    @IsString()
-    @IsNotEmpty()
-    password: string;
-
-}
