--- src/modules/user/dto/user-response.dto.ts
@@ -1,7 +1,0 @@
-export class UserResponseDto {
-
-    nickname: string;
-    accountID: string;
-    province: string;
-    city: string;
-}
